Исправляем тесты

+++ src/test/java/ru/dab/shaihulud/JsonParseTests.java
@@ -5,6 +5,8 @@ import org.json.JSONTokener;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
+import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.util.Map;
 
 public class JsonParseTests {
@@ -13,6 +15,6 @@ public class JsonParseTests {
     JSONObject obj = new JSONObject(new JSONTokener("{\"a\":1,\"b\":1.0}"));
     Map<String, Object> map = obj.toMap();
     Assertions.assertTrue(map.get("a") instanceof Integer);
-    Assertions.assertTrue(map.get("b") instanceof Double);
+    Assertions.assertTrue(map.get("b") instanceof BigDecimal);
   }
 }
